# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PriceHub API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface OkResponse {
  ok: boolean;
}

export type Role = (typeof Role)[keyof typeof Role];

export const Role = {
  buyer: "buyer",
  seller: "seller",
  developer: "developer",
  support: "support",
} as const;

export type RegisterBodyRole =
  (typeof RegisterBodyRole)[keyof typeof RegisterBodyRole];

export const RegisterBodyRole = {
  buyer: "buyer",
  seller: "seller",
} as const;

export interface RegisterBody {
  hubId: string;
  password: string;
  displayName: string;
  email?: string;
  role: RegisterBodyRole;
}

export interface LoginBody {
  hubId: string;
  password: string;
}

export interface User {
  id: number;
  hubId: string;
  displayName: string;
  /** @nullable */
  email?: string | null;
  role: Role;
  /** @nullable */
  avatarUrl?: string | null;
  banned: boolean;
  subscriptionTier: string;
  createdAt: string;
  permissions: string[];
}

export interface AuthSession {
  user: User;
}

export interface Product {
  id: number;
  sellerId: number;
  name: string;
  /** @nullable */
  description?: string | null;
  price: number;
  cost: number;
  currency: string;
  marketplace: string;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  externalUrl?: string | null;
  stock: number;
  createdAt: string;
  updatedAt: string;
}

export interface ProductCreate {
  name: string;
  description?: string;
  price: number;
  cost: number;
  currency?: string;
  marketplace: string;
  category?: string;
  imageUrl?: string;
  externalUrl?: string;
  stock?: number;
}

export interface ProductUpdate {
  name?: string;
  description?: string;
  price?: number;
  cost?: number;
  currency?: string;
  marketplace?: string;
  category?: string;
  imageUrl?: string;
  externalUrl?: string;
  stock?: number;
}

export interface MarketplaceOffer {
  marketplace: string;
  price: number;
  url: string;
  inStock: boolean;
  deliveryDays?: number;
  rating?: number;
}

export interface MarketplaceListing {
  id: number;
  name: string;
  /** @nullable */
  brand?: string | null;
  /** @nullable */
  description?: string | null;
  category: string;
  imageUrl: string;
  currency: string;
  lowestPrice: number;
  highestPrice: number;
  savingsPercent?: number;
  offers: MarketplaceOffer[];
}

export interface SalesPoint {
  date: string;
  units: number;
  revenue: number;
  profit: number;
}

export type SalesAnalyticsMarketplaceBreakdownItem = {
  marketplace: string;
  revenue: number;
  units: number;
};

export interface SalesAnalytics {
  series: SalesPoint[];
  marketplaceBreakdown: SalesAnalyticsMarketplaceBreakdownItem[];
}

export interface SellerSummary {
  revenueToday: number;
  revenueYesterday: number;
  profitToday: number;
  profitYesterday: number;
  unitsToday: number;
  unitsYesterday: number;
  productCount: number;
  conversionRate: number;
}

export interface AiInsight {
  period: string;
  headline: string;
  summary: string;
  sentiment?: string;
  recommendations: string[];
  generatedAt: string;
}

export interface AiInsightsResponse {
  yesterday: AiInsight;
  today: AiInsight;
}

export interface NewsItem {
  id: number;
  title: string;
  summary: string;
  source: string;
  category: string;
  /** @nullable */
  url?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  publishedAt: string;
}

export interface ContactUser {
  id: number;
  hubId: string;
  displayName: string;
  role: Role;
  /** @nullable */
  avatarUrl?: string | null;
}

export interface ChatThread {
  id: number;
  otherUser: ContactUser;
  /** @nullable */
  lastMessage?: string | null;
  lastMessageAt: string;
  unread: number;
}

export interface ChatMessage {
  id: number;
  chatId: number;
  senderId: number;
  senderName?: string;
  content: string;
  sentAt: string;
}

export interface ChatCreate {
  otherUserId: number;
}

export interface SendMessageBody {
  content: string;
}

export interface SubscriptionPlan {
  id: string;
  name: string;
  priceMonthly: number;
  currency: string;
  audience: string;
  tagline?: string;
  features: string[];
}

export interface UserSubscription {
  planId: string;
  status: string;
  renewsAt: string;
}

export interface SubscribeBody {
  planId: string;
}

export interface ProfileSettings {
  hubId: string;
  displayName: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  company?: string | null;
}

export interface UpdateProfileBody {
  displayName?: string;
  email?: string;
  bio?: string;
  avatarUrl?: string;
  company?: string;
}

export interface Preferences {
  theme: string;
  accent: string;
  density: string;
  language: string;
  notifications: boolean;
}

export interface PreferencesUpdate {
  theme?: string;
  accent?: string;
  density?: string;
  language?: string;
  notifications?: boolean;
}

export interface AdminUser {
  id: number;
  hubId: string;
  displayName: string;
  /** @nullable */
  email?: string | null;
  role: Role;
  banned: boolean;
  subscriptionTier?: string;
  productCount: number;
  lastActiveAt: string;
  createdAt: string;
  permissions: string[];
}

export interface AdminUserUpdate {
  displayName?: string;
  email?: string;
  role?: Role;
  banned?: boolean;
  subscriptionTier?: string;
}

export interface AdminProduct {
  id: number;
  name: string;
  sellerId: number;
  sellerName: string;
  price: number;
  cost: number;
  currency: string;
  marketplace: string;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  stock: number;
  createdAt: string;
}

export type CreateDeveloperBodyRole =
  (typeof CreateDeveloperBodyRole)[keyof typeof CreateDeveloperBodyRole];

export const CreateDeveloperBodyRole = {
  developer: "developer",
  support: "support",
} as const;

export interface CreateDeveloperBody {
  hubId: string;
  password: string;
  displayName: string;
  email?: string;
  role?: CreateDeveloperBodyRole;
  permissions: string[];
}

export interface UpdatePermissionsBody {
  permissions: string[];
}

export interface ConsoleEntry {
  id: number;
  level: string;
  source: string;
  message: string;
  createdAt: string;
}

export interface AuditEntry {
  id: number;
  actorId: number;
  actorName: string;
  actorRole: string;
  action: string;
  target: string;
  /** @nullable */
  details?: string | null;
  createdAt: string;
}

export interface ServerNode {
  id: string;
  name: string;
  region: string;
  status: string;
  cpu: number;
  memory: number;
  uptimeHours: number;
}

export interface RestartBody {
  nodeId: string;
}

export interface CodeFile {
  path: string;
  language: string;
  content: string;
  updatedAt: string;
}

export interface UpdateCodeFileBody {
  path: string;
  content: string;
}

export interface SellerHistory {
  seller: AdminUser;
  products: AdminProduct[];
  recentActions: AuditEntry[];
  salesSeries: SalesPoint[];
}

export type SearchMarketplaceParams = {
  q?: string;
  category?: string;
};