# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: PriceHub API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: registerUser
      tags: [auth]
      summary: Register a new Hub ID account
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RegisterBody" }
      responses:
        "200":
          description: Registered
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthSession" }
  /auth/login:
    post:
      operationId: loginUser
      tags: [auth]
      summary: Login with Hub ID
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginBody" }
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthSession" }
...
  /admin/users/{id}:
    patch:
      operationId: adminUpdateUser
      tags: [admin]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AdminUserUpdate" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminUser" }
    delete:
      operationId: adminDeleteUser
      tags: [admin]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OkResponse" }
...
  /admin/products/{id}:
    patch:
      operationId: adminUpdateProduct
      tags: [admin]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProductUpdate" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminProduct" }
    delete:
      operationId: adminDeleteProduct
      tags: [admin]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OkResponse" }
  /admin/devs:
    get:
      operationId: adminListDevelopers
      tags: [admin]
      responses:
        "200":
          description: Developer accounts
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/AdminUser" }
    post:
      operationId: adminCreateDeveloper
      tags: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateDeveloperBody" }
      responses:
        "200":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminUser" }
  /admin/devs/{id}/permissions:
    patch:
      operationId: adminUpdateDevPermissions
      tags: [admin]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdatePermissionsBody" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminUser" }
...
  /admin/server/restart:
    post:
      operationId: adminRestartServer
      tags: [admin]
      summary: Restart a service node
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RestartBody" }
      responses:
        "200":
          description: Restarted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OkResponse" }
...
  /admin/code:
    get:
      operationId: adminListCodeFiles
      tags: [admin]
      summary: List of code files available to the dev
      responses:
        "200":
          description: Files
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/CodeFile" }
    patch:
      operationId: adminUpdateCodeFile
      tags: [admin]
      summary: Save changes to a code file
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateCodeFileBody" }
      responses:
        "200":
          description: Saved
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CodeFile" }
  /admin/seller-history/{id}:
    get:
      operationId: adminGetSellerHistory
      tags: [admin]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Seller history
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SellerHistory" }
components:
  schemas:
...
    OkResponse:
      type: object
      properties:
        ok: { type: boolean }
      required: [ok]
    Role:
      type: string
      enum: [buyer, seller, developer, support]
    RegisterBody:
      type: object
      required: [hubId, password, displayName, role]
      properties:
        hubId: { type: string }
        password: { type: string }
        displayName: { type: string }
        email: { type: string }
        role:
          type: string
          enum: [buyer, seller]
    LoginBody:
      type: object
      required: [hubId, password]
      properties:
        hubId: { type: string }
        password: { type: string }
    AuthSession:
      type: object
      required: [user]
      properties:
        user: { $ref: "#/components/schemas/User" }
    User:
      type: object
      required: [id, hubId, displayName, role, createdAt, banned, permissions, subscriptionTier]
      properties:
        id: { type: integer }
        hubId: { type: string }
        displayName: { type: string }
        email: { type: ["string", "null"] }
        role: { $ref: "#/components/schemas/Role" }
        avatarUrl: { type: ["string", "null"] }
        banned: { type: boolean }
        subscriptionTier: { type: string }
        createdAt: { type: string, format: date-time }
        permissions:
          type: array
          items: { type: string }
...
    ProductUpdate:
      type: object
      properties:
        name: { type: string }
        description: { type: string }
        price: { type: number }
        cost: { type: number }
        currency: { type: string }
        marketplace: { type: string }
        category: { type: string }
        imageUrl: { type: string }
        externalUrl: { type: string }
        stock: { type: integer }
...
    SalesPoint:
      type: object
      required: [date, units, revenue, profit]
      properties:
        date: { type: string }
        units: { type: integer }
        revenue: { type: number }
        profit: { type: number }
...
    AdminUser:
      type: object
      required: [id, hubId, displayName, role, banned, permissions, productCount, lastActiveAt, createdAt]
      properties:
        id: { type: integer }
        hubId: { type: string }
        displayName: { type: string }
        email: { type: ["string", "null"] }
        role: { $ref: "#/components/schemas/Role" }
        banned: { type: boolean }
        subscriptionTier: { type: string }
        productCount: { type: integer }
        lastActiveAt: { type: string, format: date-time }
        createdAt: { type: string, format: date-time }
        permissions:
          type: array
          items: { type: string }
    AdminUserUpdate:
      type: object
      properties:
        displayName: { type: string }
        email: { type: string }
        role: { $ref: "#/components/schemas/Role" }
        banned: { type: boolean }
        subscriptionTier: { type: string }
    AdminProduct:
      type: object
      required: [id, name, sellerId, sellerName, price, cost, marketplace, currency, stock, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        sellerId: { type: integer }
        sellerName: { type: string }
        price: { type: number }
        cost: { type: number }
        currency: { type: string }
        marketplace: { type: string }
        category: { type: ["string", "null"] }
        imageUrl: { type: ["string", "null"] }
        stock: { type: integer }
        createdAt: { type: string, format: date-time }
    CreateDeveloperBody:
      type: object
      required: [hubId, password, displayName, permissions]
      properties:
        hubId: { type: string }
        password: { type: string }
        displayName: { type: string }
        email: { type: string }
        role:
          type: string
          enum: [developer, support]
        permissions:
          type: array
          items: { type: string }
    UpdatePermissionsBody:
      type: object
      required: [permissions]
      properties:
        permissions:
          type: array
          items: { type: string }
...
    AuditEntry:
      type: object
      required: [id, actorId, actorName, actorRole, action, target, createdAt]
      properties:
        id: { type: integer }
        actorId: { type: integer }
        actorName: { type: string }
        actorRole: { type: string }
        action: { type: string }
        target: { type: string }
        details: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
...
    RestartBody:
      type: object
      required: [nodeId]
      properties:
        nodeId: { type: string }
    CodeFile:
      type: object
      required: [path, language, content, updatedAt]
      properties:
        path: { type: string }
        language: { type: string }
        content: { type: string }
        updatedAt: { type: string, format: date-time }
    UpdateCodeFileBody:
      type: object
      required: [path, content]
      properties:
        path: { type: string }
        content: { type: string }
    SellerHistory:
      type: object
      required: [seller, products, recentActions, salesSeries]
      properties:
        seller: { $ref: "#/components/schemas/AdminUser" }
        products:
          type: array
          items: { $ref: "#/components/schemas/AdminProduct" }
        recentActions:
          type: array
          items: { $ref: "#/components/schemas/AuditEntry" }
        salesSeries:
          type: array
          items: { $ref: "#/components/schemas/SalesPoint" }